Clean up and clarify the Notifier Interaction example
The example didn't clearly demonstrate structural typing
since it explictly declared TestNotifier as implementing
Notified. It also needlessly passed the Receiver along
to the received method, even though it was never used.

This commit clarifies the use of structural typing, and
reduces the signature of received to just the message.