Build patterns nightly to get alert if it fails
Currently, a person opening a PR can be surprised by failures unrelated
to their changes. In particular, to URLs that are external to the site
no longer working and failing validation.

Worse, this means that readers of pony patterns are experiencing broken
URLs that won't be discovered until someone does an update for the
patterns and the build fails.

This commit adds a nightly job that tries to build the site and if it
fails, alerts to Zulip so we can be made aware of issues in a timely
fashion and can endeavor to fix them as soon as possible.

The workflow can also be triggered manually if need be.